python/requirements.txt
pysam>=0.21.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scanpy>=1.9.0
anndata>=0.8.0
tqdm>=4.62.0
click>=8.0.0
pyyaml>=5.4.0
